refactor #400 : Extension 분리
