+++ Outline/Outline Watch/Source/DesignSystem/Extension/ArrayExtension.swift
@@ -1,0 +1,15 @@
+//
+//  ArrayExtension.swift
+//  Outline Watch App
+//
+//  Created by hyunjun on 8/10/24.
+//
+
+import SwiftUI
+import CoreLocation
+
+extension Array where Element == Coordinate {
+    func toCLLocationCoordinates() -> [CLLocationCoordinate2D] {
+        return self.map { CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }
+    }
+}
