+++ Outline/Outline/Source/Utils/Extension/DoubleExtension.swift
